refactor(symphony): extract helper boundary
